requirements.txt
# GUARDIAN System Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Vector Database
chromadb>=0.4.22

# LLM
openai>=1.12.0
httpx>=0.26.0

# Embeddings
sentence-transformers>=2.3.0
torch>=2.1.0

# PDF Processing
pdfplumber>=0.10.3
pypdf>=3.17.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2
aiofiles>=23.2.1

# Logging
rich>=13.7.0
